npx react-native-rename ardrivingcar -b com.myviroapp.ardrivingcar
