+++ ARDrivingCar/android/app/build.gradle
@@ -126,7 +126,7 @@ android {
     compileSdkVersion rootProject.ext.compileSdkVersion
 
     defaultConfig {
-        applicationId "com.equationalapplications.arbot"
+        applicationId "com.myviroapp.ardrivingcar"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
