+++ ARDrivingCar/android/app/src/main/java/com/myviroapp/ardrivingcar/MainActivity.java
@@ -1,4 +1,4 @@
-package com.equationalapplications.arbot;
+package com.myviroapp.ardrivingcar;
 
 import com.facebook.react.ReactActivity;
 import android.os.Bundle;
@@ -11,7 +11,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "arbot";
+    return "ardrivingcar";
   }
   @Override
   protected void onCreate(Bundle savedInstanceState) {
