+++ ARDrivingCar/android/app/src/main/java/com/myviroapp/ardrivingcar/MainApplication.java
@@ -1,4 +1,4 @@
-package com.equationalapplications.arbot;
+package com.myviroapp.ardrivingcar;
 
 import android.app.Application;
 import android.content.Context;
@@ -64,7 +64,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.equationalapplications.arbot.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.myviroapp.ardrivingcar.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
