+++ ARDrivingCar/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'arbot'
+rootProject.name = 'ardrivingcar'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
 include ':react_viro', ':arcore_client', ':gvr_common', ':viro_renderer'
