+++ ARDrivingCar/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "arbot",
-  "displayName": "arbot"
+  "name": "ardrivingcar",
+  "displayName": "ardrivingcar"
 }
