+++ ARDrivingCar/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "arbot",
+  "name": "ardrivingcar",
   "version": "0.0.1",
   "private": true,
   "scripts": {
